Desearlize the JSON starting
In the createContactFromJson method I updated the code to deseralize the JSON string.
I then instantiated a Contact object and updated the values from the deseralized JSON response.
I then returned the Contact object to where the method was called from.
In the getDummyJSONUserFromId method I then updated the Contact variable to hold the new Contact object.
I then populated the DummyJSON_Id__c with the Id from the callout parameters.
Finally I upserted the contact using the DummyJSON_Id__c as the Id field to match on.